requirements.txt
tweepy==4.16.0
python-dotenv==1.1.1
python-telegram-bot==22.3